feature(two new scripts in a sort of "beta" mode): 'check_permission' is a script that checks whether or not the user has permission to read & write to either the file given as an argument, or the directory of the argument. It expects two arguments to be given: the first being a command and the second being a location. I'll probably change this to be less rigid in the future. It's very useful as an alias to editor commands like "nano", where "check_permission nano <file>" will automatically run "sudo nano <file>" if it detects that higher permissions are required. I'll want to update this script to be more applicable to any command and set of arguments as it comes in very handy.
'toggle_program' is a simple script that expects one or more arguments: the first being a command and the rest being optional arguments for that command. It first checks if the command/program is already running; if it is, it closes it instead of launching it. As an example use-case (and the reason this script exists in the first place), I love the program launcher 'wofi' but absolutely hate that it doesn't seem to have any sort of "toggle" feature. As such, hitting the keybind to open 'wofi' will just open another instance of 'wofi' instead of toggling it like 'rofi' does. This script amends that annoyance.